requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.40
reportlab==4.3.1
Pillow==11.1.0
pandas==2.2.3
numpy==2.2.4
openpyxl==3.1.5
python-dateutil==2.9.0.post0
email-validator==2.2.0
secure-smtplib==0.1.1
flask-migrate==4.1.0
python-dotenv==1.1.0
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
